+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## 2.0.0-beta.2
+
+- fix: Temporarily fudge theme token list adding button category manually
 - feat: Read Carbon version from package.json
 - chore: Change `target` option to `testOnlyTarget`